requirement.txt
appdirs==1.4.4
attrs==23.2.0
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.6
colorama==0.4.6
command-not-found==0.3
contourpy==1.0.7
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.6.1
duplicity==2.1.4
et_xmlfile==2.0.0
fasteners==0.18
Flask==3.1.0
fonttools==4.46.0
freetype-py==2.4.0
fs==2.4.16
gpg==1.18.0
html5lib==1.1
httplib2==0.20.4
idna==3.6
imageio==2.36.1
importlib-metadata==4.12.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.5
keyring==24.3.1
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mpmath==0.0.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.2
olefile==0.46
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
paramiko==2.12.0
pexpect==4.9.0
pillow==11.0.0
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pygame==2.6.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMySQL==1.1.1
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyparsing==3.1.1
pytesseract==0.3.13
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rlPyCairo==0.3.0
SciPy==1.11.4
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
sympy==1.12
systemd-python==235
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
urllib3==2.0.7
usb-creator==0.3.16
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
